using unittest for testing the node
